Refactor: Organize Go wrapper into API-specific files
The monolithic `zxc.go` file has been split into several smaller files, each
focusing on a specific ZXC API (e.g., Buffer, Block, File, Push Streaming).
This improves code organization, readability, and maintainability.

Additionally, the `zxc_io.go` file was updated to remove an arbitrary minimum
buffer size, allowing the buffer to be sized precisely based on the library's
reported needs.
