+++ wrappers/go/zxc.go
@@ -38,9 +38,6 @@ import "C"
 import (
 	"errors"
 	"fmt"
-	"os"
-	"runtime"
-	"unsafe"
 )
 
 // ============================================================================
@@ -246,752 +243,4 @@ func applyOptions(opts []Option) options {
 		fn(&o)
 	}
 	return o
-}
-
-// ============================================================================
-// Buffer API
-// ============================================================================
-
-// CompressBound returns the maximum compressed size for an input of the given
-// size. Use this to pre-allocate output buffers.
-func CompressBound(inputSize int) uint64 {
-	return uint64(C.zxc_compress_bound(C.size_t(inputSize)))
-}
-
-// Compress compresses data using the ZXC algorithm.
-//
-// Options: [WithLevel], [WithChecksum].
-//
-//	out, err := zxc.Compress(data, zxc.WithLevel(zxc.LevelCompact))
-func Compress(data []byte, opts ...Option) ([]byte, error) {
-	if len(data) == 0 {
-		return nil, ErrSrcTooSmall
-	}
-
-	o := applyOptions(opts)
-	bound := CompressBound(len(data))
-	dst := make([]byte, bound)
-
-	var copts C.zxc_compress_opts_t
-	copts.level = C.int(o.level)
-	if o.checksum {
-		copts.checksum_enabled = 1
-	}
-
-	written := C.zxc_compress(
-		unsafe.Pointer(&data[0]),
-		C.size_t(len(data)),
-		unsafe.Pointer(&dst[0]),
-		C.size_t(bound),
-		&copts,
-	)
-
-	if written < 0 {
-		return nil, errorFromCode(written)
-	}
-	if written == 0 {
-		return nil, ErrInvalidData
-	}
-
-	return dst[:int(written)], nil
-}
-
-// CompressTo compresses data into a pre-allocated output buffer.
-// Returns the number of bytes written.
-func CompressTo(data []byte, output []byte, opts ...Option) (int, error) {
-	if len(data) == 0 {
-		return 0, ErrSrcTooSmall
-	}
-	if len(output) == 0 {
-		return 0, ErrDstTooSmall
-	}
-
-	o := applyOptions(opts)
-
-	var copts C.zxc_compress_opts_t
-	copts.level = C.int(o.level)
-	if o.checksum {
-		copts.checksum_enabled = 1
-	}
-
-	written := C.zxc_compress(
-		unsafe.Pointer(&data[0]),
-		C.size_t(len(data)),
-		unsafe.Pointer(&output[0]),
-		C.size_t(len(output)),
-		&copts,
-	)
-
-	if written < 0 {
-		return 0, errorFromCode(written)
-	}
-	if written == 0 {
-		return 0, ErrInvalidData
-	}
-
-	return int(written), nil
-}
-
-// DecompressedSize returns the original uncompressed size stored in the
-// compressed data footer. Returns 0, ErrInvalidData if the data is too small
-// or invalid.
-func DecompressedSize(data []byte) (uint64, error) {
-	if len(data) == 0 {
-		return 0, ErrInvalidData
-	}
-
-	size := C.zxc_get_decompressed_size(
-		unsafe.Pointer(&data[0]),
-		C.size_t(len(data)),
-	)
-
-	if size == 0 {
-		return 0, ErrInvalidData
-	}
-	return uint64(size), nil
-}
-
-// Decompress decompresses ZXC-compressed data.
-//
-// The output size is read from the compressed data footer. For pre-allocated
-// buffers, use [DecompressTo].
-//
-// Options: [WithChecksum].
-func Decompress(data []byte, opts ...Option) ([]byte, error) {
-	if len(data) == 0 {
-		return nil, ErrInvalidData
-	}
-
-	size, err := DecompressedSize(data)
-	if err != nil {
-		return nil, err
-	}
-	if size == 0 {
-		return []byte{}, nil
-	}
-
-	o := applyOptions(opts)
-	dst := make([]byte, size)
-
-	var dopts C.zxc_decompress_opts_t
-	if o.checksum {
-		dopts.checksum_enabled = 1
-	}
-
-	written := C.zxc_decompress(
-		unsafe.Pointer(&data[0]),
-		C.size_t(len(data)),
-		unsafe.Pointer(&dst[0]),
-		C.size_t(size),
-		&dopts,
-	)
-
-	if written < 0 {
-		return nil, errorFromCode(written)
-	}
-	if uint64(written) != size {
-		return nil, ErrInvalidData
-	}
-
-	return dst[:int(written)], nil
-}
-
-// DecompressTo decompresses data into a pre-allocated output buffer.
-// Returns the number of bytes written.
-func DecompressTo(data []byte, output []byte, opts ...Option) (int, error) {
-	if len(data) == 0 {
-		return 0, ErrInvalidData
-	}
-	if len(output) == 0 {
-		return 0, ErrDstTooSmall
-	}
-
-	o := applyOptions(opts)
-
-	var dopts C.zxc_decompress_opts_t
-	if o.checksum {
-		dopts.checksum_enabled = 1
-	}
-
-	written := C.zxc_decompress(
-		unsafe.Pointer(&data[0]),
-		C.size_t(len(data)),
-		unsafe.Pointer(&output[0]),
-		C.size_t(len(output)),
-		&dopts,
-	)
-
-	if written < 0 {
-		return 0, errorFromCode(written)
-	}
-
-	return int(written), nil
-}
-
-// ============================================================================
-// Streaming API (file-based)
-// ============================================================================
-
-// CompressFile compresses src to dst using multi-threaded streaming.
-//
-// This is the recommended method for large files. It uses an asynchronous
-// pipeline with separate reader, worker, and writer threads.
-//
-// Options: [WithLevel], [WithChecksum], [WithThreads].
-//
-//	n, err := zxc.CompressFile("input.bin", "output.zxc",
-//	    zxc.WithLevel(zxc.LevelCompact), zxc.WithThreads(4))
-func CompressFile(input, output string, opts ...Option) (int64, error) {
-	o := applyOptions(opts)
-
-	fIn, err := os.Open(input)
-	if err != nil {
-		return 0, fmt.Errorf("zxc: open input: %w", err)
-	}
-	defer fIn.Close()
-
-	fOut, err := os.Create(output)
-	if err != nil {
-		return 0, fmt.Errorf("zxc: create output: %w", err)
-	}
-	defer fOut.Close()
-
-	// Duplicate file descriptors so the C FILE* owns its own fd.
-	cIn, err := dupFileRead(fIn)
-	if err != nil {
-		return 0, err
-	}
-	defer C.fclose(cIn)
-
-	cOut, err := dupFileWrite(fOut)
-	if err != nil {
-		return 0, err
-	}
-	defer C.fclose(cOut)
-
-	var copts C.zxc_compress_opts_t
-	copts.n_threads = C.int(o.threads)
-	copts.level = C.int(o.level)
-	if o.checksum {
-		copts.checksum_enabled = 1
-	}
-	if o.seekable {
-		copts.seekable = 1
-	}
-
-	result := C.zxc_stream_compress(cIn, cOut, &copts)
-	if result < 0 {
-		return 0, errorFromCode(result)
-	}
-	return int64(result), nil
-}
-
-// FileDecompressedSize reads the original uncompressed size from the footer
-// of a ZXC compressed file without performing decompression. The file is
-// opened in "rb" mode; the underlying file position is restored internally.
-func FileDecompressedSize(path string) (int64, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return 0, fmt.Errorf("zxc: open input: %w", err)
-	}
-	defer f.Close()
-
-	cf, err := dupFileRead(f)
-	if err != nil {
-		return 0, err
-	}
-	defer C.fclose(cf)
-
-	result := C.zxc_stream_get_decompressed_size(cf)
-	if result < 0 {
-		return 0, errorFromCode(result)
-	}
-	return int64(result), nil
-}
-
-// DecompressFile decompresses src to dst using multi-threaded streaming.
-//
-// Options: [WithChecksum], [WithThreads].
-//
-//	n, err := zxc.DecompressFile("compressed.zxc", "output.bin")
-func DecompressFile(input, output string, opts ...Option) (int64, error) {
-	o := applyOptions(opts)
-
-	fIn, err := os.Open(input)
-	if err != nil {
-		return 0, fmt.Errorf("zxc: open input: %w", err)
-	}
-	defer fIn.Close()
-
-	fOut, err := os.Create(output)
-	if err != nil {
-		return 0, fmt.Errorf("zxc: create output: %w", err)
-	}
-	defer fOut.Close()
-
-	cIn, err := dupFileRead(fIn)
-	if err != nil {
-		return 0, err
-	}
-	defer C.fclose(cIn)
-
-	cOut, err := dupFileWrite(fOut)
-	if err != nil {
-		return 0, err
-	}
-	defer C.fclose(cOut)
-
-	var dopts C.zxc_decompress_opts_t
-	dopts.n_threads = C.int(o.threads)
-	if o.checksum {
-		dopts.checksum_enabled = 1
-	}
-
-	result := C.zxc_stream_decompress(cIn, cOut, &dopts)
-	if result < 0 {
-		return 0, errorFromCode(result)
-	}
-	return int64(result), nil
-}
-
-// ============================================================================
-// Block API (single block, no file framing)
-// ============================================================================
-
-// CompressBlockBound returns the maximum compressed size for a single block
-// of inputSize bytes (no file framing, no EOF, no footer).
-func CompressBlockBound(inputSize int) uint64 {
-	if inputSize < 0 {
-		return 0
-	}
-	return uint64(C.zxc_compress_block_bound(C.size_t(inputSize)))
-}
-
-// DecompressBlockBound returns the minimum destination buffer size required
-// by [Cctx.DecompressBlock] for a block of uncompressedSize bytes.
-//
-// The fast decoder uses speculative wild-copy writes and needs a small tail
-// pad beyond the declared uncompressed size. Callers that cannot oversize
-// their destination buffer should use [Dctx.DecompressBlockSafe] instead.
-func DecompressBlockBound(uncompressedSize int) uint64 {
-	if uncompressedSize < 0 {
-		return 0
-	}
-	return uint64(C.zxc_decompress_block_bound(C.size_t(uncompressedSize)))
-}
-
-// EstimateCctxSize returns an accurate estimate of the memory a compression
-// context reserves when compressing a single block of srcSize bytes via
-// [Cctx.CompressBlock].
-//
-// The estimate covers all per-chunk working buffers (chain table, literals,
-// sequence/token/offset/extras buffers) plus the fixed hash tables and
-// cache-line alignment padding. It scales roughly linearly with srcSize and
-// is intended for integrators that need an accurate memory budget.
-//
-// Returns 0 if srcSize is 0 or negative.
-func EstimateCctxSize(srcSize int) uint64 {
-	if srcSize <= 0 {
-		return 0
-	}
-	return uint64(C.zxc_estimate_cctx_size(C.size_t(srcSize)))
-}
-
-// Cctx is a reusable compression context for the Block API.
-//
-// It wraps an opaque C handle that is freed by [Cctx.Close]. Create with
-// [NewCctx] and always defer a call to Close to release the native memory.
-type Cctx struct {
-	ptr *C.zxc_cctx
-}
-
-// NewCctx creates a new compression context.
-//
-// Options [WithLevel], [WithChecksum] and [WithBlockSize] are supported at
-// creation time to pre-allocate internal buffers; otherwise allocation is
-// deferred to the first call to [Cctx.CompressBlock].
-func NewCctx(opts ...Option) (*Cctx, error) {
-	o := applyOptions(opts)
-	var copts C.zxc_compress_opts_t
-	copts.level = C.int(o.level)
-	if o.checksum {
-		copts.checksum_enabled = 1
-	}
-	// Block size is optional; 0 lets the library pick the default.
-
-	ptr := C.zxc_create_cctx(&copts)
-	if ptr == nil {
-		return nil, ErrMemory
-	}
-	return &Cctx{ptr: ptr}, nil
-}
-
-// Close releases the native resources held by the context. Safe to call
-// multiple times.
-func (c *Cctx) Close() error {
-	if c == nil || c.ptr == nil {
-		return nil
-	}
-	C.zxc_free_cctx(c.ptr)
-	c.ptr = nil
-	return nil
-}
-
-// CompressBlock compresses a single block using the context. Output format
-// is [block_header(8B) + payload (+ optional checksum 4B)]. Use
-// [CompressBlockBound] to size dst.
-func (c *Cctx) CompressBlock(src, dst []byte, opts ...Option) (int, error) {
-	if c == nil || c.ptr == nil {
-		return 0, ErrNullInput
-	}
-	if len(src) == 0 {
-		return 0, ErrSrcTooSmall
-	}
-	if len(dst) == 0 {
-		return 0, ErrDstTooSmall
-	}
-
-	o := applyOptions(opts)
-	var copts C.zxc_compress_opts_t
-	copts.level = C.int(o.level)
-	if o.checksum {
-		copts.checksum_enabled = 1
-	}
-
-	n := C.zxc_compress_block(
-		c.ptr,
-		unsafe.Pointer(&src[0]),
-		C.size_t(len(src)),
-		unsafe.Pointer(&dst[0]),
-		C.size_t(len(dst)),
-		&copts,
-	)
-	if n < 0 {
-		return 0, errorFromCode(n)
-	}
-	return int(n), nil
-}
-
-// Dctx is a reusable decompression context for the Block API.
-type Dctx struct {
-	ptr *C.zxc_dctx
-}
-
-// NewDctx creates a new decompression context.
-func NewDctx() (*Dctx, error) {
-	ptr := C.zxc_create_dctx()
-	if ptr == nil {
-		return nil, ErrMemory
-	}
-	return &Dctx{ptr: ptr}, nil
-}
-
-// Close releases the native resources held by the context.
-func (d *Dctx) Close() error {
-	if d == nil || d.ptr == nil {
-		return nil
-	}
-	C.zxc_free_dctx(d.ptr)
-	d.ptr = nil
-	return nil
-}
-
-// DecompressBlock decompresses a single block produced by
-// [Cctx.CompressBlock].
-//
-// dst should be at least [DecompressBlockBound](uncompressedSize) to enable
-// the fast decode path. For a strictly-sized destination buffer use
-// [Dctx.DecompressBlockSafe] instead.
-func (d *Dctx) DecompressBlock(src, dst []byte, opts ...Option) (int, error) {
-	if d == nil || d.ptr == nil {
-		return 0, ErrNullInput
-	}
-	if len(src) == 0 {
-		return 0, ErrSrcTooSmall
-	}
-	if len(dst) == 0 {
-		return 0, ErrDstTooSmall
-	}
-
-	o := applyOptions(opts)
-	var dopts C.zxc_decompress_opts_t
-	if o.checksum {
-		dopts.checksum_enabled = 1
-	}
-
-	n := C.zxc_decompress_block(
-		d.ptr,
-		unsafe.Pointer(&src[0]),
-		C.size_t(len(src)),
-		unsafe.Pointer(&dst[0]),
-		C.size_t(len(dst)),
-		&dopts,
-	)
-	if n < 0 {
-		return 0, errorFromCode(n)
-	}
-	return int(n), nil
-}
-
-// DecompressBlockSafe is a strict-sized variant of [Dctx.DecompressBlock]:
-// it accepts a destination buffer sized exactly to the uncompressed length,
-// with no tail-pad required. Slightly slower than the fast path; output is
-// bit-identical.
-func (d *Dctx) DecompressBlockSafe(src, dst []byte, opts ...Option) (int, error) {
-	if d == nil || d.ptr == nil {
-		return 0, ErrNullInput
-	}
-	if len(src) == 0 {
-		return 0, ErrSrcTooSmall
-	}
-	if len(dst) == 0 {
-		return 0, ErrDstTooSmall
-	}
-
-	o := applyOptions(opts)
-	var dopts C.zxc_decompress_opts_t
-	if o.checksum {
-		dopts.checksum_enabled = 1
-	}
-
-	n := C.zxc_decompress_block_safe(
-		d.ptr,
-		unsafe.Pointer(&src[0]),
-		C.size_t(len(src)),
-		unsafe.Pointer(&dst[0]),
-		C.size_t(len(dst)),
-		&dopts,
-	)
-	if n < 0 {
-		return 0, errorFromCode(n)
-	}
-	return int(n), nil
-}
-
-// ============================================================================
-// Push Streaming API (single-threaded, caller-driven)
-// ============================================================================
-
-// CStream is a push-based, single-threaded compression stream — the Go
-// counterpart of the C [zxc_cstream]. Use it to integrate ZXC into event
-// loops, callback-driven libraries, or non-blocking network protocols where
-// the [CompressFile] FILE*-based pipeline is not appropriate.
-//
-// CStream is not safe for concurrent use; one goroutine per stream.
-type CStream struct {
-	ptr *C.zxc_cstream
-}
-
-// NewCStream creates a push compression stream.
-//
-// Honoured options: [WithLevel], [WithChecksum]. [WithThreads] and
-// [WithSeekable] are ignored (the push API is single-threaded and never
-// emits a seek table).
-func NewCStream(opts ...Option) (*CStream, error) {
-	o := applyOptions(opts)
-	var copts C.zxc_compress_opts_t
-	copts.level = C.int(o.level)
-	if o.checksum {
-		copts.checksum_enabled = 1
-	}
-	ptr := C.zxc_cstream_create(&copts)
-	if ptr == nil {
-		return nil, ErrMemory
-	}
-	return &CStream{ptr: ptr}, nil
-}
-
-// Close releases the stream and all internal buffers. Safe to call multiple
-// times.
-func (c *CStream) Close() error {
-	if c == nil || c.ptr == nil {
-		return nil
-	}
-	C.zxc_cstream_free(c.ptr)
-	c.ptr = nil
-	return nil
-}
-
-// InSize returns the suggested input chunk size for best throughput.
-func (c *CStream) InSize() int {
-	if c == nil || c.ptr == nil {
-		return 0
-	}
-	return int(C.zxc_cstream_in_size(c.ptr))
-}
-
-// OutSize returns the suggested output chunk size to never trigger a partial
-// drain.
-func (c *CStream) OutSize() int {
-	if c == nil || c.ptr == nil {
-		return 0
-	}
-	return int(C.zxc_cstream_out_size(c.ptr))
-}
-
-// Compress pushes bytes from in into the stream and writes compressed bytes
-// to out. Returns:
-//
-//   - consumed: bytes read from in.
-//   - produced: bytes written into out.
-//   - pending:  bytes still staged inside the stream (drain out and call
-//     again with the same or new in to continue).
-//
-// A return of (consumed, produced, 0, nil) means in was fully consumed and
-// no compressed bytes remain pending. Calling with an empty in is valid
-// (drain-only mode).
-func (c *CStream) Compress(out, in []byte) (consumed, produced int, pending int64, err error) {
-	if c == nil || c.ptr == nil {
-		return 0, 0, 0, ErrNullInput
-	}
-
-	var pinner runtime.Pinner
-	defer pinner.Unpin()
-
-	var inBuf C.zxc_inbuf_t
-	if len(in) > 0 {
-		pinner.Pin(&in[0])
-		inBuf.src = unsafe.Pointer(&in[0])
-	}
-	inBuf.size = C.size_t(len(in))
-
-	var outBuf C.zxc_outbuf_t
-	if len(out) > 0 {
-		pinner.Pin(&out[0])
-		outBuf.dst = unsafe.Pointer(&out[0])
-	}
-	outBuf.size = C.size_t(len(out))
-
-	r := C.zxc_cstream_compress(c.ptr, &outBuf, &inBuf)
-	if r < 0 {
-		return 0, 0, 0, errorFromCode(r)
-	}
-	return int(inBuf.pos), int(outBuf.pos), int64(r), nil
-}
-
-// End finalises the stream by flushing the residual block, then writing the
-// EOF block and the file footer to out. Call repeatedly while pending > 0,
-// draining out between calls.
-//
-// After End returns (produced, 0, nil), the stream is in DONE state; further
-// calls return ErrNullInput. Always call [CStream.Close] to release the
-// native resources.
-func (c *CStream) End(out []byte) (produced int, pending int64, err error) {
-	if c == nil || c.ptr == nil {
-		return 0, 0, ErrNullInput
-	}
-
-	var pinner runtime.Pinner
-	defer pinner.Unpin()
-
-	var outBuf C.zxc_outbuf_t
-	if len(out) > 0 {
-		pinner.Pin(&out[0])
-		outBuf.dst = unsafe.Pointer(&out[0])
-	}
-	outBuf.size = C.size_t(len(out))
-
-	r := C.zxc_cstream_end(c.ptr, &outBuf)
-	if r < 0 {
-		return 0, 0, errorFromCode(r)
-	}
-	return int(outBuf.pos), int64(r), nil
-}
-
-// DStream is a push-based, single-threaded decompression stream — the Go
-// counterpart of the C [zxc_dstream].
-type DStream struct {
-	ptr *C.zxc_dstream
-}
-
-// NewDStream creates a push decompression stream.
-//
-// Honoured options: [WithChecksum]. [WithThreads] is ignored.
-func NewDStream(opts ...Option) (*DStream, error) {
-	o := applyOptions(opts)
-	var dopts C.zxc_decompress_opts_t
-	if o.checksum {
-		dopts.checksum_enabled = 1
-	}
-	ptr := C.zxc_dstream_create(&dopts)
-	if ptr == nil {
-		return nil, ErrMemory
-	}
-	return &DStream{ptr: ptr}, nil
-}
-
-// Close releases the stream and all internal buffers. Safe to call multiple
-// times.
-func (d *DStream) Close() error {
-	if d == nil || d.ptr == nil {
-		return nil
-	}
-	C.zxc_dstream_free(d.ptr)
-	d.ptr = nil
-	return nil
-}
-
-// InSize returns the suggested input chunk size for the decompressor.
-func (d *DStream) InSize() int {
-	if d == nil || d.ptr == nil {
-		return 0
-	}
-	return int(C.zxc_dstream_in_size(d.ptr))
-}
-
-// OutSize returns the suggested output chunk size for the decompressor.
-func (d *DStream) OutSize() int {
-	if d == nil || d.ptr == nil {
-		return 0
-	}
-	return int(C.zxc_dstream_out_size(d.ptr))
-}
-
-// Finished reports whether the decoder has reached and validated the file
-// footer. Useful to detect truncated streams: if the input source is
-// drained and Finished returns false, the stream ended prematurely.
-func (d *DStream) Finished() bool {
-	if d == nil || d.ptr == nil {
-		return false
-	}
-	return C.zxc_dstream_finished(d.ptr) != 0
-}
-
-// Decompress pushes compressed bytes from in into the stream and writes
-// decompressed bytes to out. Returns:
-//
-//   - consumed: bytes read from in.
-//   - produced: bytes written into out.
-//
-// A return of (0, 0, nil) when in is non-empty means no progress could be
-// made: either the parser is waiting for more input (feed more), or the
-// stream has reached DONE state (any trailing bytes in in are ignored).
-// Use [DStream.Finished] to disambiguate.
-func (d *DStream) Decompress(out, in []byte) (consumed, produced int, err error) {
-	if d == nil || d.ptr == nil {
-		return 0, 0, ErrNullInput
-	}
-
-	var pinner runtime.Pinner
-	defer pinner.Unpin()
-
-	var inBuf C.zxc_inbuf_t
-	if len(in) > 0 {
-		pinner.Pin(&in[0])
-		inBuf.src = unsafe.Pointer(&in[0])
-	}
-	inBuf.size = C.size_t(len(in))
-
-	var outBuf C.zxc_outbuf_t
-	if len(out) > 0 {
-		pinner.Pin(&out[0])
-		outBuf.dst = unsafe.Pointer(&out[0])
-	}
-	outBuf.size = C.size_t(len(out))
-
-	r := C.zxc_dstream_decompress(d.ptr, &outBuf, &inBuf)
-	if r < 0 {
-		return 0, 0, errorFromCode(r)
-	}
-	return int(inBuf.pos), int(outBuf.pos), nil
 }
